+++ src/UKHO.S100PermitService.API/Controllers/PermitController.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using UKHO.S100PermitService.Common.Events;
+using UKHO.S100PermitService.Common.IO;
 using UKHO.S100PermitService.Common.Models;
 using UKHO.S100PermitService.Common.Services;
-using UKHO.S100PermitService.Common.Utilities;
 
 namespace UKHO.S100PermitService.API.Controllers
 {
