+++ tests/UKHO.S100PermitService.API.UnitTests/Controller/PermitControllerTests.cs
@@ -6,8 +6,8 @@ using Microsoft.Extensions.Logging;
 using NUnit.Framework;
 using UKHO.S100PermitService.API.Controllers;
 using UKHO.S100PermitService.Common.Events;
+using UKHO.S100PermitService.Common.IO;
 using UKHO.S100PermitService.Common.Services;
-using UKHO.S100PermitService.Common.Utilities;
 
 namespace UKHO.S100PermitService.API.UnitTests.Controller
 {
